Traverse JSDoc nodes if present
This change has two open questions that manifest as ugly code:

1. The jsDoc property is not exposed in the Typescript API because the
intended interface is `getJSDocTags`, which does a lot more work to
provide a list of tags that are collected directly from a node, or any
one of several parent nodes. I use type assertions to any to access the
property without checking.
2. The jsDoc property is of type JSDoc[], not Node, so it can't be
traversed directly. I decided to add the discovered JSDoc
nodes as children before the actual children, since that's where JSDoc
appears syntactically. But it's odd that that each tag appears as an
individual child instead of appearing under some synthetic jsdoc node.
Unfortunately, I don't know tsquery well enough to know if such a
synthetic node is possible, or a good idea.

Both problems boil down to "JSDoc is stored on a side-band to the AST",
which leads to problems parsing and representing it as part of the AST.